DSServer/requirements.txt
aiohttp==3.6.2
asgiref==3.2.5
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
color==0.1
colorama==0.4.3
comtypes==1.1.7
COVID19Py==0.3.0
decorator==4.4.2
discord.py==1.3.4
Django==3.0.4
docutils==0.16
enum34==1.1.10
future==0.18.2
geocoder==1.38.1
geojson==2.5.0
get==2019.4.13
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.9
keyboard==0.13.5
Kivy-Garden==0.1.4
mouse==0.7.1
MouseInfo==0.1.3
multidict==4.7.5
numpy==1.18.2
opencv-python==4.2.0.34
pathtools==0.1.2
Pillow==7.1.1
pip==20.1.1
post==2019.4.13
power==1.4
public==2019.4.13
pyasn1-modules==0.2.8
PyAutoGUI==0.9.50
pycparser==2.20
pygame==1.9.6
PyGetWindow==0.0.8
Pygments==2.6.1
PyMsgBox==1.0.7
PyNaCl==1.3.0
pynput==1.6.8
pyopenssl==19.1.0
pyowm==2.10.0
pyperclip==1.8.0
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
pyTelegramBotAPI==3.6.7
pyttsx3==2.87
PyTweening==1.0.3
pytz==2019.3
pywin32==227
query-string==2019.4.13
ratelim==0.1.6
request==2019.4.13
requests==2.7.0
secure-smtplib==0.1.1
service-identity==18.1.0
setuptools==46.1.3
six==1.14.0
soupsieve==2.0
SpeechRecognition==3.8.1
sqlparse==0.3.1
telebot==0.0.3
telegrambotapi==0.3.2
tkinter-nav==0.0.5
tkintertable==1.3.2
urllib3==1.25.8
watchdog==0.10.2
websockets==8.1
wheel==0.34.2
yarl==1.4.2